snapshots: move snapshot directory (v3)
Originally we kept snapshots under /var/lib/lxcsnaps.  If a
separate btrfs is mounted at /var/lib/lxc, then we can't
make btrfs snapshots under /var/lib/lxcsnaps.

This patch moves the default directory to /var/lib/lxc/c/snaps.
If /var/lib/lxcsnaps already exists, then use that.

If we are deleting a container which has snapshots, we currently
will delete the container itself and its rootfs, but not its
snapshots.  This could be confusing for the user, and there is
no option to c->destroy() to ask for different behavior.  So
currently a user would have to delete all snapshots first, then
delete the container.  Ideas for better handling this would be
welcome, but we don't want to change the current api, so while
adding a new c->destroy_full() would be ok, adding a flags
argument to c->destroy(c, flags) is not.

lxclock: use ".$lxcname" for container lock files

that way we can use /run/lock/lxc/$lxcpath/$lxcname/snaps as a
directory when locking snapshots without having to worry about
/run/lock//lxc/$lxcpath/$lxcname being a file.

snapshot: update paths in testcase

TODO: either change how we handle snapshots when deleting a
container, or better clean up in src/tests/snapshot.c, so that
lxc-test-snapshot can succeed twice in a row.  (Currently it
cannot, bc snapxxx1 has two snapshots and snapxxx3 has one.

+++ src/lxc/utils.c
@@ -46,7 +46,8 @@
 
 lxc_log_define(lxc_utils, lxc);
 
-static int _recursive_rmdir_onedev(char *dirname, dev_t pdev)
+static int _recursive_rmdir_onedev(char *dirname, dev_t pdev,
+				   const char *exclude, int level)
 {
 	struct dirent dirent, *direntp;
 	DIR *dir;
@@ -70,6 +71,9 @@ static int _recursive_rmdir_onedev(char *dirname, dev_t pdev)
 		    !strcmp(direntp->d_name, ".."))
 			continue;
 
+		if (!level && exclude && !strcmp(direntp->d_name, exclude))
+			continue;
+
 		rc = snprintf(pathname, MAXPATHLEN, "%s/%s", dirname, direntp->d_name);
 		if (rc < 0 || rc >= MAXPATHLEN) {
 			ERROR("pathname too long");
@@ -85,7 +89,7 @@ static int _recursive_rmdir_onedev(char *dirname, dev_t pdev)
 		if (mystat.st_dev != pdev)
 			continue;
 		if (S_ISDIR(mystat.st_mode)) {
-			if (_recursive_rmdir_onedev(pathname, pdev) < 0)
+			if (_recursive_rmdir_onedev(pathname, pdev, exclude, level+1) < 0)
 				failed=1;
 		} else {
 			if (unlink(pathname) < 0) {
@@ -95,9 +99,11 @@ static int _recursive_rmdir_onedev(char *dirname, dev_t pdev)
 		}
 	}
 
-	if (rmdir(dirname) < 0) {
-		ERROR("%s: failed to delete %s", __func__, dirname);
-		failed=1;
+	if (!exclude || level != 0) {
+		if (rmdir(dirname) < 0) {
+			ERROR("%s: failed to delete %s", __func__, dirname);
+			failed=1;
+		}
 	}
 
 	ret = closedir(dir);
@@ -110,7 +116,7 @@ static int _recursive_rmdir_onedev(char *dirname, dev_t pdev)
 }
 
 /* returns 0 on success, -1 if there were any failures */
-extern int lxc_rmdir_onedev(char *path)
+extern int lxc_rmdir_onedev(char *path, const char *exclude)
 {
 	struct stat mystat;
 
@@ -119,7 +125,7 @@ extern int lxc_rmdir_onedev(char *path)
 		return -1;
 	}
 
-	return _recursive_rmdir_onedev(path, mystat.st_dev);
+	return _recursive_rmdir_onedev(path, mystat.st_dev, exclude, 0);
 }
 
 static int mount_fs(const char *source, const char *target, const char *type)
